Fix imports: ProjectUser entity moved to its own folder
